refactor: extract strip_to_plain_text into utils::text
Move the HTML-to-plaintext stripper from the handler module to a reusable
utility so DB and model layers can import it without a layering smell.
