+++ src/handlers/pages/search_text.rs
@@ -1,85 +1,7 @@
 //! Plain-text extraction, snippet building, and match highlighting for the
 //! SSR search results page. Pure string functions — no request state.
 
-fn strip_to_plain_text(raw: &str) -> String {
-    let mut out = String::with_capacity(raw.len());
-    let mut in_tag = false;
-    let mut skip_until: Option<&'static str> = None;
-    let mut last_space = true;
-    let bytes = raw.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        if let Some(end_tag) = skip_until {
-            if let Some(pos) = raw[i..].to_ascii_lowercase().find(end_tag) {
-                i += pos + end_tag.len();
-                skip_until = None;
-                in_tag = false;
-                if !last_space {
-                    out.push(' ');
-                    last_space = true;
-                }
-                continue;
-            } else {
-                break;
-            }
-        }
-        let ch = bytes[i] as char;
-        match ch {
-            '<' => {
-                let lower = raw[i..].to_ascii_lowercase();
-                if lower.starts_with("<script") {
-                    skip_until = Some("</script>");
-                    i += 1;
-                    continue;
-                }
-                if lower.starts_with("<style") {
-                    skip_until = Some("</style>");
-                    i += 1;
-                    continue;
-                }
-                if lower.starts_with("<!--") {
-                    if let Some(pos) = raw[i + 4..].find("-->") {
-                        i += 4 + pos + 3;
-                        if !last_space {
-                            out.push(' ');
-                            last_space = true;
-                        }
-                        continue;
-                    } else {
-                        break;
-                    }
-                }
-                in_tag = true;
-                i += 1;
-            }
-            '>' if in_tag => {
-                in_tag = false;
-                if !last_space {
-                    out.push(' ');
-                    last_space = true;
-                }
-                i += 1;
-            }
-            _ if in_tag => {
-                i += 1;
-            }
-            c if c.is_whitespace() => {
-                if !last_space {
-                    out.push(' ');
-                    last_space = true;
-                }
-                i += 1;
-            }
-            _ => {
-                let ch_len = raw[i..].chars().next().map_or(1, |c| c.len_utf8());
-                out.push_str(&raw[i..i + ch_len]);
-                last_space = false;
-                i += ch_len;
-            }
-        }
-    }
-    out.trim().to_string()
-}
+use crate::utils::text::strip_to_plain_text;
 
 /// Build a query-aware snippet: returns a `max_chars`-wide window centered on
 /// the first case-insensitive match of `query` in the plain-text content, with
