+++ src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod datetime;
+pub mod text;
 pub mod url_validation;
